@@ -65,8 +65,8 @@
     "rollup-plugin-replace": "^2.0.0",
     "rollup-plugin-uglify": "^6.0.0",
     "uglify-es": "^3.1.6",
-    "vue": "^2.5.16",
-    "vue-property-decorator": "^8.2.2"
+    "vue": "^3.0.0",
+    "vue-property-decorator": "^10.0.0-rc.3"
   },
   "resolutions": {
     "js-yaml": "^3.13.1"
