@@ -1,6 +1,6 @@
 /* eslint no-unused-vars: 0 */
 
-import Vue, { AsyncComponent } from 'vue'
+import { ComponentOptions, AsyncComponentOptions, App } from 'vue'
 import { ApolloClient } from '@apollo/client/core'
 import {
   VueApolloComponentOptions,
@@ -8,19 +8,20 @@ import {
   ErrorHandler,
 } from './options'
 
-export type VueApolloComponent<V extends Vue = Vue> = VueApolloComponentOptions<V> | typeof Vue | AsyncComponent
+export type VueApolloComponent<V extends ComponentOptions = ComponentOptions> = VueApolloComponentOptions<V> | AsyncComponentOptions
 
-export interface ApolloProviderOptions {
+export interface ApolloProviderOptions<TCacheShape = any> {
   defaultClient: ApolloClient<TCacheShape>
-  defaultOptions?: VueApolloComponentOptions<Vue>
+  defaultOptions?: VueApolloComponentOptions<ComponentOptions>
   clients?: { [key: string]: ApolloClient<TCacheShape> }
   watchLoading?: WatchLoading
   errorHandler?: ErrorHandler
   prefetch?: boolean
 }
+
 export class ApolloProvider<TCacheShape=any> {
-  provide: (key?: string) => this
-  constructor (options: ApolloProviderOptions)
+  constructor (options: ApolloProviderOptions<TCacheShape>)
+  install (app: App): void
 
   clients: { [key: string]: ApolloClient<TCacheShape> }
   defaultClient: ApolloClient<TCacheShape>
