@@ -1,12 +1,14 @@
+/* eslint-disable import/no-duplicates */
+
 declare module '*.gql' {
-  import { DocumentNode, DocumentNode } from 'graphql'
+  import { DocumentNode } from 'graphql'
 
   const content: DocumentNode
   export default content
-
 }
 
 declare module '*.graphql' {
+  import { DocumentNode } from 'graphql'
 
   const content: DocumentNode
   export default content
