@@ -1,24 +1,26 @@
-import Vue from 'vue'
-import { CombinedVueInstance } from 'vue/types/vue'
 import { DollarApollo } from './vue-apollo'
 import { VueApolloComponentOptions } from './options'
 import { ApolloProvider } from './apollo-provider'
 
-type DataDef<Data, Props, V> = Data | ((this: Readonly<Props> & V) => Data)
-
-declare module 'vue/types/options' {
-  interface ComponentOptions<V extends Vue, Data, Methods, Computed, PropsDef, Props> {
+declare module '@vue/runtime-core' {
+  interface ComponentOptionsBase<
+    Props,
+    RawBindings,
+    D,
+    C extends ComputedOptions,
+    M extends MethodOptions,
+    Mixin extends ComponentOptionsMixin,
+    Extends extends ComponentOptionsMixin,
+    E extends EmitsOptions,
+    EE extends string = string,
+    // eslint-disable-next-line @typescript-eslint/ban-types
+    Defaults = {}
+  > {
     apolloProvider?: ApolloProvider
-    apollo?: VueApolloComponentOptions<
-    Data extends DataDef<infer D, any, any>
-      ? CombinedVueInstance<V, D, Methods, Computed, Props>
-      : CombinedVueInstance<V, Data, Methods, Computed, Props>
-    >
+    apollo?: VueApolloComponentOptions<CreateComponentPublicInstance<Props, RawBindings, D, C, M, Mixin, Extends, E, Props, Defaults, false>>
   }
-}
 
-declare module 'vue/types/vue' {
-  interface Vue {
+  interface ComponentCustomProperties {
     $apolloProvider: ApolloProvider
     $apollo: DollarApollo<this>
   }
